ported trigger tests to jest
